requirements.txt
Flask==1.1.2
Flask-Cors==3.0.10
flask-restplus==0.13.0
Werkzeug==0.16.1
waitress==2.0.0
psycopg2==2.8.6
opencv-python==3.4.8.29
numpy==1.21.2
Pillow==6.2.1
torch==1.9.1
torchvision==0.10.1
